SIIM19_Pneumothorax_Segmentation_2nd_solution/requirements.txt
addict==2.2.1
albumentations==0.3.0
apex==0.1
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
Cython==0.29.10
decorator==4.4.0
efficientnet-pytorch==0.3.0
efficientunet-pytorch==0.0.5
idna==2.8
imageio==2.5.0
imgaug==0.2.6
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
mmcv==0.2.8
mock==3.0.5
munch==2.3.2
networkx==2.3
numexpr==2.6.9
numpy==1.16.4
opencv-python==4.1.0.25
opencv-python-headless==4.1.0.25
pandas==0.24.2
pathlib==1.0.1
pbkdf2==1.3
Pillow==6.0.0
pretrainedmodels==0.7.4
protobuf==3.8.0
pycocotools==2.0.0
pydensecrf==1.0rc3
pyparsing==2.4.0
python-dateutil==2.8.0
pytorchcv==0.0.48
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.2
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
segmentation-models-pytorch==0.0.1
six==1.12.0
some-package==0.1
tables==3.5.2
tensorboardX==1.7
terminaltables==3.1.0
torch==1.1.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.1
urllib3==1.25.3
visdom==0.1.8.8
websocket-client==0.56.0
